+++ packages/vue/src/use-assistant.ui.test.tsx
@@ -1,8 +1,8 @@
 import { formatAssistantStreamPart } from '@ai-sdk/ui-utils';
 import {
-  mockFetchDataStream,
-  mockFetchDataStreamWithGenerator,
-} from '@ai-sdk/ui-utils/test';
+  createTestServer,
+  TestResponseController,
+} from '@ai-sdk/provider-utils/test';
 import '@testing-library/jest-dom/vitest';
 import userEvent from '@testing-library/user-event';
 import { cleanup, findByText, render, screen } from '@testing-library/vue';
@@ -10,13 +10,16 @@ import TestChatAssistantStreamComponent from './TestChatAssistantStreamComponent
 import TestChatAssistantThreadChangeComponent from './TestChatAssistantThreadChangeComponent.vue';
 import { setupTestComponent } from './setup-test-component';
 
+const server = createTestServer({
+  '/api/assistant': {},
+});
+
 describe('stream data stream', () => {
   setupTestComponent(TestChatAssistantStreamComponent);
 
   it('should show streamed response', async () => {
-    // Mock the fetch data stream
-    const { requestBody } = mockFetchDataStream({
-      url: 'https://example.com/api/assistant',
+    server.urls['/api/assistant'].response = {
+      type: 'stream-chunks',
       chunks: [
         // Format the stream part
         formatAssistantStreamPart('assistant_control_data', {
@@ -33,7 +36,7 @@ describe('stream data stream', () => {
         '0:", world"\n',
         '0:"."\n',
       ],
-    });
+    };
 
     // Click the button
     await userEvent.click(screen.getByTestId('do-append'));
@@ -50,60 +53,44 @@ describe('stream data stream', () => {
       'AI: Hello, world.',
     );
 
-    expect(await requestBody).toStrictEqual(
-      JSON.stringify({
-        message: 'hi',
-        threadId: null,
-      }),
-    );
+    expect(await server.calls[0].requestBody).toStrictEqual({
+      message: 'hi',
+      threadId: null,
+    });
   });
 
   describe('loading state', () => {
     it('should show loading state', async () => {
-      let finishGeneration: ((value?: unknown) => void) | undefined;
+      const controller = new TestResponseController();
+      server.urls['/api/assistant'].response = {
+        type: 'controlled-stream',
+        controller,
+      };
 
-      const finishGenerationPromise = new Promise(resolve => {
-        finishGeneration = resolve;
-      });
-
-      // Mock the fetch data stream with generator
-      mockFetchDataStreamWithGenerator({
-        url: 'https://example.com/api/assistant',
-        chunkGenerator: (async function* generate() {
-          const encoder = new TextEncoder();
-
-          yield encoder.encode(
-            formatAssistantStreamPart('assistant_control_data', {
-              threadId: 't0',
-              messageId: 'm1',
-            }),
-          );
-
-          yield encoder.encode(
-            formatAssistantStreamPart('assistant_message', {
-              id: 'm1',
-              role: 'assistant',
-              content: [{ type: 'text', text: { value: '' } }],
-            }),
-          );
-
-          yield encoder.encode('0:"Hello"\n');
-
-          await finishGenerationPromise;
-        })(),
-      });
-
-      // Click the button
       await userEvent.click(screen.getByTestId('do-append'));
 
       // Find the loading element and expect it to be in progress
       await screen.findByTestId('status');
       expect(screen.getByTestId('status')).toHaveTextContent('in_progress');
 
-      // Resolve the finishGenerationPromise
-      finishGeneration?.();
+      controller.write(
+        formatAssistantStreamPart('assistant_control_data', {
+          threadId: 't0',
+          messageId: 'm1',
+        }),
+      );
 
-      // Find the loading element and expect it to be awaiting a message
+      controller.write(
+        formatAssistantStreamPart('assistant_message', {
+          id: 'm1',
+          role: 'assistant',
+          content: [{ type: 'text', text: { value: '' } }],
+        }),
+      );
+
+      controller.write('0:"Hello"\n');
+      controller.close();
+
       await findByText(await screen.findByTestId('status'), 'awaiting_message');
       expect(screen.getByTestId('status')).toHaveTextContent(
         'awaiting_message',
@@ -128,8 +115,8 @@ describe('Thread management', () => {
   });
 
   it('should show streamed response', async () => {
-    const { requestBody } = mockFetchDataStream({
-      url: 'https://example.com/api/assistant',
+    server.urls['/api/assistant'].response = {
+      type: 'stream-chunks',
       chunks: [
         formatAssistantStreamPart('assistant_control_data', {
           threadId: 't0',
@@ -146,7 +133,7 @@ describe('Thread management', () => {
         '0:" world"\n',
         '0:"."\n',
       ],
-    });
+    };
 
     await userEvent.click(screen.getByTestId('do-append'));
 
@@ -160,12 +147,10 @@ describe('Thread management', () => {
       'AI: Hello, world.',
     );
 
-    expect(await requestBody).toStrictEqual(
-      JSON.stringify({
-        message: 'hi',
-        threadId: null,
-      }),
-    );
+    expect(await server.calls[0].requestBody).toStrictEqual({
+      message: 'hi',
+      threadId: null,
+    });
   });
 
   it('should switch to new thread on setting undefined threadId', async () => {
@@ -174,8 +159,8 @@ describe('Thread management', () => {
     expect(screen.queryByTestId('message-0')).toBeNull();
     expect(screen.queryByTestId('message-1')).toBeNull();
 
-    const { requestBody } = mockFetchDataStream({
-      url: 'https://example.com/api/assistant',
+    server.urls['/api/assistant'].response = {
+      type: 'stream-chunks',
       chunks: [
         formatAssistantStreamPart('assistant_control_data', {
           threadId: 't1',
@@ -186,13 +171,12 @@ describe('Thread management', () => {
           role: 'assistant',
           content: [{ type: 'text', text: { value: '' } }],
         }),
-        // text parts:
         '0:"Hello"\n',
         '0:","\n',
         '0:" world"\n',
         '0:"."\n',
       ],
-    });
+    };
 
     await userEvent.click(screen.getByTestId('do-append'));
 
@@ -207,12 +191,10 @@ describe('Thread management', () => {
     );
 
     // check that correct information was sent to the server:
-    expect(await requestBody).toStrictEqual(
-      JSON.stringify({
-        message: 'hi',
-        threadId: null,
-      }),
-    );
+    expect(await server.calls[0].requestBody).toStrictEqual({
+      message: 'hi',
+      threadId: null,
+    });
   });
 
   it('should switch to thread on setting previously created threadId', async () => {
@@ -221,8 +203,8 @@ describe('Thread management', () => {
     expect(screen.queryByTestId('message-0')).toBeNull();
     expect(screen.queryByTestId('message-1')).toBeNull();
 
-    const { requestBody } = mockFetchDataStream({
-      url: 'https://example.com/api/assistant',
+    server.urls['/api/assistant'].response = {
+      type: 'stream-chunks',
       chunks: [
         formatAssistantStreamPart('assistant_control_data', {
           threadId: 't3',
@@ -239,7 +221,7 @@ describe('Thread management', () => {
         '0:" world"\n',
         '0:"."\n',
       ],
-    });
+    };
 
     await userEvent.click(screen.getByTestId('do-append'));
 
@@ -253,12 +235,9 @@ describe('Thread management', () => {
       'AI: Hello, world.',
     );
 
-    // check that correct information was sent to the server:
-    expect(await requestBody).toStrictEqual(
-      JSON.stringify({
-        message: 'hi',
-        threadId: 't3',
-      }),
-    );
+    expect(await server.calls[0].requestBody).toStrictEqual({
+      message: 'hi',
+      threadId: 't3',
+    });
   });
 });