+++ packages/anthropic/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @ai-sdk/anthropic
 
+## 1.2.6
+
+### Patch Changes
+
+- Updated dependencies [2c19b9a]
+  - @ai-sdk/provider-utils@2.2.4
+
+## 1.2.5
+
+### Patch Changes
+
+- 292f543: fix (provider/google-vertex): fix anthropic support for image urls in messages
+
 ## 1.2.4
 
 ### Patch Changes