+++ packages/cohere/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/cohere
 
+## 1.2.5
+
+### Patch Changes
+
+- Updated dependencies [2c19b9a]
+  - @ai-sdk/provider-utils@2.2.4
+
 ## 1.2.4
 
 ### Patch Changes