+++ packages/groq/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @ai-sdk/groq
 
+## 1.2.5
+
+### Patch Changes
+
+- 8fdc8fe: feat (provider/groq): add llama 4 model
+
+## 1.2.4
+
+### Patch Changes
+
+- Updated dependencies [2c19b9a]
+  - @ai-sdk/provider-utils@2.2.4
+
 ## 1.2.3
 
 ### Patch Changes