+++ packages/openai-compatible/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @ai-sdk/openai-compatible
 
+## 0.2.6
+
+### Patch Changes
+
+- Updated dependencies [2c19b9a]
+  - @ai-sdk/provider-utils@2.2.4
+
+## 0.2.5
+
+### Patch Changes
+
+- d186cca: feat (provider/openai-compatible): add additional token usage metrics
+
 ## 0.2.4
 
 ### Patch Changes